+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KMG2P2EMP7398N7TJ3S066MR"
+        "01KMG5HC4S9M9NGX0XYHSVR0GA"
     )
     r"""The unique identifier of the data source"""