+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KMG2P2EK3J7EQ80EM5ZM0PR2"
+        "01KMG5HC4RSF02B1Z0J6TYKEFG"
     )
     r"""The unique identifier of the data source"""