+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KMG2P2EP52KA1MPVM027X28Z"
+        "01KMG5HC4WNDRKSCHXK3ESPY15"
     )
     r"""The unique identifier of the data source"""