+++ packages/orq-rc/RELEASES.md
@@ -6569,3 +6569,13 @@ Based on:
 - [python v4.7.0-rc.16] packages/orq-rc
 ### Releases
 - [PyPI v4.7.0-rc.16] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.16 - packages/orq-rc
+
+## 2026-03-24 14:59:16
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.759.1 (2.869.10) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.7.0-rc.17] packages/orq-rc
+### Releases
+- [PyPI v4.7.0-rc.17] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.17 - packages/orq-rc